Move src/utils/deep-readonly.ts -> src/utils/types/deep-readonly.ts
